Refactoring command line parser.

+++ src/include/udjat/tools/string.h
@@ -253,7 +253,7 @@
 
 		/// @brief Test if the string contains one of the elements of a list.
 		/// @return Index of the matched content (-1 if not found).
-		size_t select(const char *value, ...) __attribute__ ((sentinel));
+		int select(const char *value, ...) __attribute__ ((sentinel));
 
 		/// @brief Insert global expander.
 		/// @param method String expander method (returns 'true' if the value was parsed).
